Huge improvement
completed the Game class (at least the functions to be used so far);
wrote class Engine (so many facade methods);

added a new Component: Canvas;
added a new interface: Paintable (and some implements or adapters) which is used by Canvas;

wrote the reference counter stystem for ImageBufferData, added the facade methods to Engine either;

fought against the big,  if not the biggest, bug in paint method (the bound intersection detection and cliping while it works on an animated Component);
crushed the bug;
found that it's impossible to crush it with java.awt.Graphics, but who cares;

refined some constructors in Button;
yet another slightly syntax improvement;

finally have time for a shower;

OK, now it only remains the class Data (old DataBase), and replacing the old text format with new one, and tests. Then celebrate for the first step completement (you finally can call it version 0.1.0)!